ADD deserialize middleware
