+++ controllers/user.controller.js
@@ -30,42 +30,26 @@ async function postLogin(req, res, next) {
 }
 
 async function getUserByToken(req, res, next) {
-  const { authorization } = req.cookies;
+  const { user } = req;
 
-  // NOTE: deserialize에서 걸러준다면, 없어도 되는 케이스일 것 같다
-  if (!authorization) {
-    next(createError(401, "authorization is not exist."));
+  if (!user) {
+    next(createError(401, "user is not exist."));
     return;
   }
 
-  const token = authorization.split("bearer ")[1];
-
-  try {
-    const { id } = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
-    const user = await User.findById(id)?.lean();
-
-    if (!user) {
-      next(createError(400, "user is not exist."));
-      return;
-    }
-
-    res.json({ ok: true, data: user });
-  } catch (err) {
-    if (err instanceof jwt.JsonWebTokenError) {
-      next(createError(401, "authorization is not correct."));
-      return;
-    }
-
-    next(err);
-  }
+  res.json({ ok: true, data: user });
 }
 
-// NOTE: deserialize에서 req.user === null인 케이스를 걸렀다는 전제로 작성
 async function deleteUser(req, res, next) {
-  const { _id } = req.user;
+  const { user } = req;
+
+  if (!user) {
+    next(createError(401, "authorization is invalid."));
+    return;
+  }
 
   try {
-    const user = await User.deleteOne({ _id });
+    await User.deleteOne({ _id: user._id });
 
     res.json({ ok: true, data: user });
   } catch (err) {
@@ -76,8 +60,7 @@ async function deleteUser(req, res, next) {
 async function getUserById(req, res, next) {
   const { id } = req.params;
 
-  // NOTE: id의 형식이 _id인지를 확인해주는 부분인데, 현재 db의 data기준 오류남. 어디가 잘못된것이지?
-  if (!(id instanceof mongoose.Types.ObjectId)) {
+  if (!(mongoose.Types.ObjectId.isValid(id))) {
     next(createError(400, "id of param is invalid."));
     return;
   }
