+++ routes/user.js
@@ -1,11 +1,12 @@
 const router = require("express").Router();
 
 const userController = require("../controllers/user.controller");
+const deserialize = require("../middlewares/deserialize");
 
 // TODO: rename
 router.post("/", userController.postLogin);
-router.get("/", userController.getUserByToken);
-router.delete("/", userController.deleteUser);
+router.get("/", deserialize, userController.getUserByToken);
+router.delete("/", deserialize, userController.deleteUser);
 
 router.get("/:id", userController.getUserById);
